+++ tests/Comment/CommentTest.php
@@ -49,8 +49,8 @@ final class CommentTest extends TestCase
             . ' * Comments' . "\n"
             . ' *//* Hell */@import url("some/url.css") screen;'
             . '/* Number 4 *//* Number 5 */.foo,#bar{'
-            . '/* Number 6 */background-color:#000;}@media screen{'
-            . '/** Number 10 **/#foo.bar{/** Number 10b **/position:absolute;}}',
+            . '/* Number 6 */background-color:#000}@media screen{'
+            . '/** Number 10 **/#foo.bar{/** Number 10b **/position:absolute}}',
             $cssDocument->render(OutputFormat::createCompact()->setRenderComments(true))
         );
     }
@@ -76,8 +76,8 @@ final class CommentTest extends TestCase
 ', $css->render(OutputFormat::createPretty()->setRenderComments(false)));
         self::assertSame(
             '@import url("some/url.css") screen;'
-            . '.foo,#bar{background-color:#000;}'
-            . '@media screen{#foo.bar{position:absolute;}}',
+            . '.foo,#bar{background-color:#000}'
+            . '@media screen{#foo.bar{position:absolute}}',
             $css->render(OutputFormat::createCompact())
         );
     }