+++ tests/OutputFormatTest.php
@@ -69,8 +69,8 @@ EOT;
     public function compact(): void
     {
         self::assertSame(
-            '.main,.test{font:italic normal bold 16px/1.2 "Helvetica",Verdana,sans-serif;background:white;}'
-            . '@media screen{.main{background-size:100% 100%;font-size:1.3em;background-color:#fff;}}',
+            '.main,.test{font:italic normal bold 16px/1.2 "Helvetica",Verdana,sans-serif;background:white}'
+            . '@media screen{.main{background-size:100% 100%;font-size:1.3em;background-color:#fff}}',
             $this->document->render(OutputFormat::createCompact())
         );
     }