+++ src/Property/Charset.php
@@ -99,6 +99,8 @@ class Charset implements AtRule, Positionable
 
     /**
      * @return string
+     *
+     * @deprecated since V8.9.0, will be removed in version 9.0.0. Use the specific getter instead.
      */
     public function atRuleArgs()
     {