+++ CHANGELOG.md
@@ -42,6 +42,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Deprecated
 
+- Deprecate `::atRuleArgs()` (#1191)
 - Support for PHP < 7.2 is deprecated; version 9.0 will require PHP 7.2 or later
   (#1264)
 - Passing a `string` or `null` to `RuleSet::removeRule()` is deprecated